Add Tenant and ApplicationUser models. Add CatalogDbContext for tenant management. Refactoring.
