+++ src/PolicyManagement.Application/Extensions/ApplicationServiceExtensions.cs
@@ -3,27 +3,26 @@ using Microsoft.Extensions.Hosting;
 using Serilog;
 using Serilog.Events;
 
-namespace PolicyManagement.Application.Extensions
+namespace PolicyManagement.Application.Extensions;
+
+public static class ApplicationServiceExtensions
 {
-    public static class ApplicationServiceExtensions
+    public static IServiceCollection AddApplicationServices(this IServiceCollection services)
     {
-        public static IServiceCollection AddApplicationServices(this IServiceCollection services)
-        {
-            // Register application services here
+        // Register application services here
 
-            return services;
-        }
+        return services;
+    }
 
-        public static IHostBuilder UseSerilogConfiguration(this IHostBuilder host)
-        {
-            host.UseSerilog((context, services, configuration) => configuration
-                .ReadFrom.Configuration(context.Configuration)
-                .ReadFrom.Services(services)
-                .Enrich.FromLogContext()
-                .MinimumLevel.Override("Microsoft", LogEventLevel.Warning)
-                .MinimumLevel.Override("Microsoft.AspNetCore", LogEventLevel.Warning));
+    public static IHostBuilder UseSerilogConfiguration(this IHostBuilder host)
+    {
+        host.UseSerilog((context, services, configuration) => configuration
+            .ReadFrom.Configuration(context.Configuration)
+            .ReadFrom.Services(services)
+            .Enrich.FromLogContext()
+            .MinimumLevel.Override("Microsoft", LogEventLevel.Warning)
+            .MinimumLevel.Override("Microsoft.AspNetCore", LogEventLevel.Warning));
 
-            return host;
-        }
+        return host;
     }
 }
