+++ src/PolicyManagement.Domain/Entities/Identity/ApplicationRole.cs
@@ -1,0 +1,14 @@
+using Microsoft.AspNetCore.Identity;
+
+namespace PolicyManagement.Domain.Entities.Identity;
+
+public class ApplicationRole : IdentityRole
+{
+    public ApplicationRole() : base()
+    {
+    }
+
+    public ApplicationRole(string roleName) : base(roleName)
+    {
+    }
+}
