--- src/PolicyManagement.Persistence/Models/Identity/ApplicationRole.cs
@@ -1,11 +1,0 @@
-using Microsoft.AspNetCore.Identity;
-
-namespace PolicyManagement.Persistence.Models.Identity
-{
-    public class ApplicationRole : IdentityRole
-    {
-        public string? Description { get; set; }
-        public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
-        public DateTime? LastModifiedAt { get; set; }
-    }
-}
