--- src/PolicyManagement.Persistence/Models/Identity/ApplicationUser.cs
@@ -1,27 +1,0 @@
-using Microsoft.AspNetCore.Identity;
-using System;
-using System.ComponentModel.DataAnnotations;
-using System.ComponentModel.DataAnnotations.Schema;
-using PolicyManagement.Persistence.Models.Catalog;
-
-namespace PolicyManagement.Persistence.Models.Identity
-{
-    public class ApplicationUser : IdentityUser
-    {
-        [MaxLength(100)]
-        public string? FirstName { get; set; }
-        
-        [MaxLength(100)]
-        public string? LastName { get; set; }
-        
-        [Required]
-        public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
-        
-        public DateTime? LastModifiedAt { get; set; }
-        
-        [ForeignKey("Tenant")]
-        public Guid? TenantId { get; set; }
-        
-        public virtual Tenant? Tenant { get; set; }
-    }
-}
