+++ src/PolicyManagementApp.Api/Models/ApiModels/ErrorResponseModel.cs
@@ -1,8 +1,7 @@
-namespace PolicyManagementApp.Api.Models.ApiModels
+namespace PolicyManagementApp.Api.Models.ApiModels;
+
+public class ErrorResponseModel
 {
-    public class ErrorResponseModel
-    {
-        public string Error { get; set; } = "An error occurred.";
-        public int StatusCode { get; set; }
-    }
-} 
+    public string Error { get; set; } = "An error occurred.";
+    public int StatusCode { get; set; }
+}
