+++ src/PolicyManagementApp.Api/appsettings.json
@@ -1,7 +1,13 @@
 {
   "ConnectionStrings": {
-    "CatalogConnection": "Data Source=WALID-HOME\\WALIDHOME;Initial Catalog=catalog;Persist Security Info=True;User ID=sa;Password=sa;Trust Server Certificate=True"
+    "Catalog": "Data Source=WALID-HOME\\WALIDHOME;Initial Catalog=Catalog;Persist Security Info=True;User ID=sa;Password=sa;Trust Server Certificate=True",
+    "Tenant1Db": "Data Source=WALID-HOME\\WALIDHOME;Initial Catalog=Tenant1Db;Persist Security Info=True;User ID=sa;Password=sa;Trust Server Certificate=True",
+    "Tenant2Db": "Data Source=WALID-HOME\\WALIDHOME;Initial Catalog=Tenant2Db;Persist Security Info=True;User ID=sa;Password=sa;Trust Server Certificate=True"
   },
+  "FeatureFlags": {
+    "EnableSeedData": true
+  },
+  "AllowedHosts": "*",
   "Serilog": {
     "Using": [
       "Serilog.Sinks.Console",
@@ -39,10 +45,5 @@
       "Application": "PolicyManagementApp",
       "Environment": "Development"
     }
-  },
-  "Demo": {
-    "Description": "Set to false in production environments. When true, default tenants and users will be seeded for demonstration purposes.",
-    "SeedData": false
-  },
-  "AllowedHosts": "*"
+  }
 }
